back-end/requirements.txt
certifi==2019.3.9
chardet==3.0.4
dataclasses==0.6
Django==2.2
django-cors-headers==2.5.2
djangorestframework==3.9.2
djongo==1.2.32
idna==2.8
oauthlib==3.0.1
pymongo==3.7.2
PySocks==1.6.8
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
slackclient==1.3.1
sqlparse==0.2.4
tweepy==3.7.0
urllib3==1.24.1
websocket-client==0.54.0
